+++ AutoLayoutTutorial/Assets.xcassets/venom-7.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "data" : [
+    {
+      "idiom" : "universal",
+      "filename" : "venom-7.jp2"
+    }
+  ]
+}
