+++ AutoLayoutTutorial/ProfileCell.swift
@@ -1,0 +1,45 @@
+//
+//  ProfileCell.swift
+//  AutoLayoutTutorial
+//
+//  Created by Stephen Dowless on 4/8/19.
+//  Copyright © 2019 Stephan Dowless. All rights reserved.
+//
+
+import UIKit
+
+class ProfileCell: UICollectionViewCell {
+    
+    // MARK: - Properties
+    
+    var image: UIImage? {
+        didSet {
+            guard let image = image else { return }
+            cellImageView.image = image
+        }
+    }
+    
+    let cellImageView: UIImageView = {
+        let iv = UIImageView()
+        iv.image = #imageLiteral(resourceName: "venom")
+        iv.contentMode = .scaleAspectFill
+        iv.clipsToBounds = true
+        return iv
+    }()
+    
+    // MARK: - Lifecycle
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        addSubview(cellImageView)
+        cellImageView.anchor(top: topAnchor, left: leftAnchor, bottom: bottomAnchor, right: rightAnchor)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    
+    
+}
